Move `storage::[async_]{get_child_nodes,node_exists,node_exists,listable}` to `node` module
Remove `[async_]discover_nodes`
