+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add `impl TryInto<StorePrefix> for &NodePath`
    - Removes `TryFrom<&NodePath> for StorePrefix`
  - **Breaking**: Move `extract_byte_ranges_read[_seek]` from `array::codec` to `storage::byte_range`
+ - **Breaking**: Move `storage::[async_]{get_child_nodes,node_exists,node_exists,listable}` to `node` module
 
 ### Removed
  - **Breaking**: Remove `array::NonZeroError`, use `std::num::TryFromIntError` instead
@@ -59,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `[async_]create_{array, group}`
     - `[async_]erase_{chunk,metadata}`, `[async_]{retrieve,store}_chunk`
     - `[async_]retrieve_partial_values`
+    - `[async_]discover_nodes`
  - **Breaking**: Remove `Default` implementation for `Metadata{Convert,Erase}Version`
     - Explicitly use `global_config()` instead
  - **Breaking**: Remove `array::UnsafeCellSlice`
