+++ src/storage/storage_async.rs
@@ -1,22 +1,11 @@
 use std::sync::Arc;
 
-use async_recursion::async_recursion;
-
 use futures::{StreamExt, TryStreamExt};
 use itertools::Itertools;
 
-use crate::{
-    array::{ArrayMetadata, ArrayMetadataV2},
-    group::GroupMetadata,
-    metadata::{v2::GroupMetadataV2, v3::NodeMetadataV3, NodeMetadata},
-    node::{Node, NodePath},
-    storage::meta_key_v3,
-};
-
 use super::{
-    byte_range::ByteRange, meta_key_v2_array, meta_key_v2_attributes, meta_key_v2_group,
-    AsyncBytes, MaybeAsyncBytes, StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue,
-    StoreKeys, StoreKeysPrefixes, StorePrefix, StorePrefixes,
+    byte_range::ByteRange, AsyncBytes, MaybeAsyncBytes, StorageError, StoreKey, StoreKeyRange,
+    StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix, StorePrefixes,
 };
 
 /// Async readable storage traits.
@@ -305,105 +294,7 @@ impl<T> AsyncReadableWritableListableStorageTraits for T where
 {
 }
 
-async fn get_metadata_v3<
-    TStorage: ?Sized + AsyncReadableStorageTraits + AsyncListableStorageTraits,
->(
-    storage: &Arc<TStorage>,
-    prefix: &StorePrefix,
-) -> Result<Option<NodeMetadata>, StorageError> {
-    let key: StoreKey = meta_key_v3(&prefix.try_into()?);
-    match storage.get(&key).await? {
-        Some(metadata) => {
-            let metadata: NodeMetadataV3 = serde_json::from_slice(&metadata)
-                .map_err(|err| StorageError::InvalidMetadata(key, err.to_string()))?;
-            Ok(Some(match metadata {
-                NodeMetadataV3::Array(array) => NodeMetadata::Array(ArrayMetadata::V3(array)),
-                NodeMetadataV3::Group(group) => NodeMetadata::Group(GroupMetadata::V3(group)),
-            }))
-        }
-        None => Ok(None),
-    }
-}
-
-async fn get_metadata_v2<
-    TStorage: ?Sized + AsyncReadableStorageTraits + AsyncListableStorageTraits,
->(
-    storage: &Arc<TStorage>,
-    prefix: &StorePrefix,
-) -> Result<Option<NodeMetadata>, StorageError> {
-    let node_path = prefix.try_into()?;
-    let attributes_key = meta_key_v2_attributes(&node_path);
-
-    // Try array
-    let key_array: StoreKey = meta_key_v2_array(&node_path);
-    if let Some(metadata) = storage.get(&key_array).await? {
-        let mut metadata: ArrayMetadataV2 = serde_json::from_slice(&metadata)
-            .map_err(|err| StorageError::InvalidMetadata(key_array, err.to_string()))?;
-        let attributes = storage.get(&attributes_key).await?;
-        if let Some(attributes) = attributes {
-            let attributes: serde_json::Map<String, serde_json::Value> =
-                serde_json::from_slice(&attributes).map_err(|err| {
-                    StorageError::InvalidMetadata(attributes_key, err.to_string())
-                })?;
-            metadata.attributes = attributes;
-        }
-        return Ok(Some(NodeMetadata::Array(ArrayMetadata::V2(metadata))));
-    }
-
-    // Try group
-    let key_group: StoreKey = meta_key_v2_group(&node_path);
-    if let Some(metadata) = storage.get(&key_group).await? {
-        let mut metadata: GroupMetadataV2 = serde_json::from_slice(&metadata)
-            .map_err(|err| StorageError::InvalidMetadata(key_group, err.to_string()))?;
-        let attributes = storage.get(&attributes_key).await?;
-        if let Some(attributes) = attributes {
-            let attributes: serde_json::Map<String, serde_json::Value> =
-                serde_json::from_slice(&attributes).map_err(|err| {
-                    StorageError::InvalidMetadata(attributes_key, err.to_string())
-                })?;
-            metadata.attributes = attributes;
-        }
-        return Ok(Some(NodeMetadata::Group(GroupMetadata::V2(metadata))));
-    }
-
-    Ok(None)
-}
-
-/// Asynchronously get the child nodes.
-///
-/// # Errors
-/// Returns a [`StorageError`] if there is an underlying error with the store.
-#[async_recursion]
-pub async fn async_get_child_nodes<TStorage>(
-    storage: &Arc<TStorage>,
-    path: &NodePath,
-) -> Result<Vec<Node>, StorageError>
-where
-    TStorage: ?Sized + AsyncReadableStorageTraits + AsyncListableStorageTraits,
-{
-    let prefixes = async_discover_children(storage, path).await?;
-    let mut nodes: Vec<Node> = Vec::new();
-    // TODO: Asynchronously get metadata of all prefixes
-    for prefix in &prefixes {
-        let mut child_metadata = get_metadata_v3(storage, prefix).await?;
-        if child_metadata.is_none() {
-            child_metadata = get_metadata_v2(storage, prefix).await?;
-        }
-        let Some(child_metadata) = child_metadata else {
-            return Err(StorageError::MissingMetadata(prefix.clone()));
-        };
-
-        let path: NodePath = prefix.try_into()?;
-        let children = match child_metadata {
-            NodeMetadata::Array(_) => Vec::default(),
-            NodeMetadata::Group(_) => async_get_child_nodes(storage, &path).await?,
-        };
-        nodes.push(Node::new_with_metadata(path, child_metadata, children));
-    }
-    Ok(nodes)
-}
-
-/// Asynchronously discover the children of a node.
+/// Asynchronously discover the children of a store prefix.
 ///
 /// # Errors
 /// Returns a [`StorageError`] if there is an underlying error with the store.
@@ -411,11 +302,10 @@ pub async fn async_discover_children<
     TStorage: ?Sized + AsyncReadableStorageTraits + AsyncListableStorageTraits,
 >(
     storage: &Arc<TStorage>,
-    path: &NodePath,
+    prefix: &StorePrefix,
 ) -> Result<StorePrefixes, StorageError> {
-    let prefix: StorePrefix = path.try_into()?;
     let children: Result<Vec<_>, _> = storage
-        .list_dir(&prefix)
+        .list_dir(prefix)
         .await?
         .prefixes()
         .iter()
@@ -423,60 +313,4 @@ pub async fn async_discover_children<
         .map(|v| StorePrefix::new(v.as_str()))
         .collect();
     Ok(children?)
-}
-
-/// Asynchronously discover all nodes.
-///
-/// # Errors
-/// Returns a [`StorageError`] if there is an underlying error with the store.
-///
-pub async fn async_discover_nodes(
-    storage: &dyn AsyncListableStorageTraits,
-) -> Result<StoreKeys, StorageError> {
-    storage.list_prefix(&"".try_into()?).await
-}
-
-/// Asynchronously erase a node (group or array) and all of its children.
-///
-/// Returns true if the node existed and was removed.
-///
-/// # Errors
-/// Returns a [`StorageError`] if there is an underlying error with the store.
-pub async fn async_erase_node(
-    storage: &dyn AsyncWritableStorageTraits,
-    path: &NodePath,
-) -> Result<(), StorageError> {
-    let prefix = path.try_into()?;
-    storage.erase_prefix(&prefix).await
-}
-
-/// Asynchronously check if a node exists.
-///
-/// # Errors
-/// Returns a [`StorageError`] if there is an underlying error with the store.
-pub async fn async_node_exists<
-    TStorage: ?Sized + AsyncReadableStorageTraits + AsyncListableStorageTraits,
->(
-    storage: &Arc<TStorage>,
-    path: &NodePath,
-) -> Result<bool, StorageError> {
-    Ok(storage.get(&meta_key_v3(path)).await?.is_some()
-        || storage.get(&meta_key_v2_array(path)).await?.is_some()
-        || storage.get(&meta_key_v2_group(path)).await?.is_some())
-}
-
-/// Asynchronously check if a node exists.
-///
-/// # Errors
-/// Returns a [`StorageError`] if there is an underlying error with the store.
-pub async fn async_node_exists_listable<TStorage: ?Sized + AsyncListableStorageTraits>(
-    storage: &Arc<TStorage>,
-    path: &NodePath,
-) -> Result<bool, StorageError> {
-    let prefix: StorePrefix = path.try_into()?;
-    storage.list_prefix(&prefix).await.map(|keys| {
-        keys.contains(&meta_key_v3(path))
-            | keys.contains(&meta_key_v2_array(path))
-            | keys.contains(&meta_key_v2_group(path))
-    })
 }
