+++ src/storage/store/store_async.rs
@@ -9,9 +9,8 @@ mod test_util {
     use std::error::Error;
 
     use crate::storage::{
-        async_discover_nodes, byte_range::ByteRange, AsyncListableStorageTraits,
-        AsyncReadableStorageTraits, AsyncWritableStorageTraits, StoreKeyRange, StoreKeyStartValue,
-        StorePrefix,
+        byte_range::ByteRange, AsyncListableStorageTraits, AsyncReadableStorageTraits,
+        AsyncWritableStorageTraits, StoreKeyRange, StoreKeyStartValue, StorePrefix,
     };
 
     /// Create a store with the following data
@@ -145,18 +144,6 @@ mod test_util {
 
         assert_eq!(
             store.list_prefix(&"".try_into()?).await?,
-            &[
-                "a/b".try_into()?,
-                "a/c".try_into()?,
-                "a/d/e".try_into()?,
-                "a/f/g".try_into()?,
-                "a/f/h".try_into()?,
-                "i/j/k".try_into()?
-            ]
-        );
-
-        assert_eq!(
-            async_discover_nodes(store).await?,
             &[
                 "a/b".try_into()?,
                 "a/c".try_into()?,
