+++ src/storage/store/store_sync.rs
@@ -12,8 +12,8 @@ mod test_util {
     use std::error::Error;
 
     use crate::storage::{
-        byte_range::ByteRange, discover_nodes, ListableStorageTraits, ReadableStorageTraits,
-        StoreKeyRange, StoreKeyStartValue, StorePrefix, WritableStorageTraits,
+        byte_range::ByteRange, ListableStorageTraits, ReadableStorageTraits, StoreKeyRange,
+        StoreKeyStartValue, StorePrefix, WritableStorageTraits,
     };
 
     /// Create a store with the following data
@@ -118,18 +118,6 @@ mod test_util {
 
         assert_eq!(
             store.list_prefix(&"".try_into()?)?,
-            &[
-                "a/b".try_into()?,
-                "a/c".try_into()?,
-                "a/d/e".try_into()?,
-                "a/f/g".try_into()?,
-                "a/f/h".try_into()?,
-                "i/j/k".try_into()?
-            ]
-        );
-
-        assert_eq!(
-            discover_nodes(store)?,
             &[
                 "a/b".try_into()?,
                 "a/c".try_into()?,
